feat: add GetWithDefault and GetReviewPrompt methods to config manager
